+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Changes
 
 - Bump the injected `cider-nrepl` to [0.32](https://github.com/clojure-emacs/cider-nrepl/blob/master/CHANGELOG.md).
+- Improve `nrepl-dict` error reporting.
 
 ## 1.7.0 (2023-03-23)